Add Int4.cpp.
